Rename overlooked package
