+++ src/main/kotlin/no/nb/mlt/wls/infrastructure/config/WLSServiceConfig.kt
@@ -7,7 +7,7 @@ import no.nb.mlt.wls.domain.ports.outbound.TransactionPort
 import no.nb.mlt.wls.infrastructure.callbacks.InventoryNotifierAdapter
 import no.nb.mlt.wls.infrastructure.repositories.item.ItemRepositoryMongoAdapter
 import no.nb.mlt.wls.infrastructure.repositories.order.MongoOrderRepositoryAdapter
-import no.nb.mlt.wls.infrastructure.repositories.outbox.MongoStorageMessageRepositoryAdapter
+import no.nb.mlt.wls.infrastructure.repositories.storageMessage.MongoStorageMessageRepositoryAdapter
 import no.nb.mlt.wls.infrastructure.synq.SynqAdapter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
